pkg/sequence/parse: renames the package from parser
