+++ pkg/sequence/parse/testdata/1.txt
